+++ src/apyfixedarray.h
@@ -269,9 +269,11 @@ public:
         nb::ndarray<nb::numpy, uint8_t>>
     to_bits(bool numpy = false) const;
 
+    //! Create an N-dimensional array containing bit-patterns.
     template <typename NB_ARRAY_TYPE, typename INT_TYPE>
     nb::ndarray<NB_ARRAY_TYPE, INT_TYPE> to_bits_ndarray() const;
 
+    //! Create a nested Python list containing bit-patterns as Python integers.
     nb::list to_bits_python_recursive_descent(
         std::size_t dim, APyBuffer<mp_limb_t>::vector_type::const_iterator& it
     ) const;